requirements_spark.txt
requests
boto3
python-dotenv
delta-spark
pytest